// frontend/src/types/course.types.ts
export interface CourseCardData{
id?:number;
slug:string;
title:string;
subTitle?: string;
thumbnail:string;
difficulty:string;
duration:number;
instructorName:string;
isInWishList:boolean;
isEnrolled?: boolean;
}
// ...
  interface Instructor {
    fullName: string;
    specialization: string | null;
    description: string | null;
    otherCourses: CourseCardData[];
  }
  
  export interface FormattedCourse {
    id: number;
    title: string;
    subtitle: string | null;
    description: string;
    difficulty: string;
    isPublished:boolean;
    slug: string;
    introVideo: string;
    duration: number;
    createdAt: Date;
    isEnrolled:boolean;
    whatYouWillLearn: []; // Could be specified more precisely based on actual structure
    requirements: []; // Could be specified more precisely based on actual structure
    categories: []; // Could be specified more precisely based on actual structure
    isInWishList: boolean;
    instructor: Instructor;
    enrollmentsCount: number;
    modules: {
      id: number;
      title: string;
      duration: number;
      orderPosition: number;
      lessons: Lesson[];
    }[];

  }